requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
cryptography>=41.0.0
pytz>=2023.3
tabulate>=0.9.0
rich>=13.0.0
